+++ spec/directions/bake_notes/bake_unclassified_notes_spec.rb
@@ -22,10 +22,33 @@ RSpec.describe Kitchen::Directions::BakeUnclassifiedNotes do
     )
   end
 
+  let(:numbered_exercise_within_note) do
+    book_containing(html:
+      one_chapter_with_one_page_containing(
+        <<~HTML
+          <div data-type="note" id="unclassifiednote">
+            <div data-type="title" id="titleId">note title</div>
+            <p>this is a note</p>
+            <div data-type="exercise" id="3360">
+              <div data-type="problem" id="504">
+                <ul>
+                  <li>What do you need to know to perform this analysis at the very minimum?</li>
+                </ul>
+              </div>
+            </div>
+          </div>
+        HTML
+      )
+    )
+  end
+
   before do
     stub_locales({
       'notes': {
         'foo': 'Bar'
+      },
+      'exercises': {
+        'exercise': 'Exercise'
       }
     })
   end
@@ -54,5 +77,34 @@ RSpec.describe Kitchen::Directions::BakeUnclassifiedNotes do
         </div>
       HTML
     )
+  end
+
+  context 'when unclassified notes have numbered exercise within' do
+    it 'bakes' do
+      described_class.v1(book: numbered_exercise_within_note, bake_exercises: true)
+      expect(numbered_exercise_within_note.body.pages.first.notes).to match_normalized_html(
+        <<~HTML
+          <div data-type="note" id="unclassifiednote">
+            <h3 class="os-title" data-type="title">
+              <span class="os-title-label" data-type="" id="titleId">note title</span>
+            </h3>
+            <div class="os-note-body">
+              <p>this is a note</p>
+              <div data-type="exercise" id="3360">
+                <div data-type="problem" id="504">
+                  <span class="os-title-label">Exercise </span>
+                  <span class="os-number">1</span>
+                  <div class="os-problem-container">
+                    <ul>
+                      <li>What do you need to know to perform this analysis at the very minimum?</li>
+                    </ul>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        HTML
+      )
+    end
   end
 end